Fix load() args
